Fix handling of multiple completion scopes
When the YAML front matter contains no "completion-scope", the
expression `[front_matter["completion-scope"]]` evaluates to `[nil]`,
which is truthy, so the rest of the chain is not evaluated and the
snippet is assigned `completion_scopes` of `[nil]`.

As it happens, `[nil]` is serialised as `<array><data>BAgw</data></array>`.

When a code snippet containing that "completion scope" is present,
Xcode crashes as soon as I try to enter any text in a place where
a snippet might be admissible.

This change replaces the chain of `||` operators with a `case`
block.  This leads to some ugly repetition in the resulting code,
but at least it doesn't crash Xcode. ;-)